feat: move db into separate package
