+++ db/db.go
@@ -1,0 +1,5 @@
+package db
+
+import "gorm.io/gorm"
+
+var db *gorm.DB
